+++ lib/core.ts
