+++ lib/global.d.ts
