+++ lib/physics.ts
