+++ lib/text.ts
