+++ pages/_app.tsx
@@ -5,7 +5,7 @@ import "@fontsource/inter/600.css"
 import "@fontsource/inter/700.css"
 import "@fontsource/inter/800.css"
 import "@fontsource/inter/900.css"
-import Theme from '@styles/Theme'
+import Theme from '@/styles/Theme'
 
 export default function App({ Component, pageProps }) {
   return (
