+++ tsconfig.json
@@ -1,8 +1,8 @@
 {
-  "include": ["src", "pages"], "exclude": ["node_modules", "**/node_modules/*"],
+  "include": ["pages", "components", "lib", "styles"],
+  "exclude": ["node_modules", "**/node_modules/*"],
   "compilerOptions": {
     "jsx": "preserve",
-    "baseUrl": "src",
     "module": "commonjs",
     "allowSyntheticDefaultImports": true,
     "target": "es5",
@@ -16,11 +16,11 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
+    "baseUrl": ".",
     "paths": {
-      "@root/*": ["../*"],
-      "@styles/*": ["styles/*"],
-      "@components/*": ["components/*"],
-      "@utils/*": ["utils/*"]
+      "@/styles/*": ["styles/*"],
+      "@/components/*": ["components/*"],
+      "@/lib/*": ["lib/*"]
     }
   }
 }
